requirements.txt
web3>=6.0.0
schedule>=1.2.0
eth-account>=0.9.0
requests>=2.28.0
python-telegram-bot>=20.0 